# docs/openapi/pn-mandate-internal-v1.yaml
openapi: 3.0.1
info:
  title: PN Mandate u-service
  description: Documentation APIs v1.0
  termsOfService: https://termofservice.it
  x-api-id: api-internal-mandate
  x-summary: 'Piattaforma Notifiche: API gestione deleghe'
  version: '1.0.0'
  contact:
    email: pn@pagopa.it
  license:
    name: Licenza di PN
    url: 'https://da-definire/'
servers:
  - url: https://api.pn.pagopa.it
    description: Generated server url
paths:
  /status:
    get:
      summary: healthCheck path
      description: health check path per verificare lo stato del micro servizio
      tags:
        - HealthCheck
      operationId: status
      responses:
        '200':
          description: Ok
        '500':
          description: Internal Server Error
  /mandate-private/api/v1/mandates-by-internaldelegate/{internaluserId}:                          # NO EXTERNAL
    get:                                                                                          # NO EXTERNAL
      tags:                                                                                       # NO EXTERNAL
        - mandate-private-service                                                                 # NO EXTERNAL
      summary: Return only active mandates for delegate, optionally filtered by single mandateId  # NO EXTERNAL
      operationId: listMandatesByDelegate                                                         # NO EXTERNAL
      parameters:
        - name: x-pagopa-pn-cx-type
          in: query
          description: cx-type (PF, PG, PA)
          required: true
          schema:
            $ref: '#/components/schemas/cxTypeAuthFleet'                                         # NO EXTERNAL
        - name: internaluserId                                                                    # NO EXTERNAL
          in: path                                                                                # NO EXTERNAL
          required: true                                                                          # NO EXTERNAL
          schema:                                                                                 # NO EXTERNAL
            type: string                                                                          # NO EXTERNAL
        - name: mandateId                                                                         # NO EXTERNAL
          in: query                                                                               # NO EXTERNAL
          schema:                                                                                 # NO EXTERNAL
            type: string                                                                          # NO EXTERNAL
        - name: x-pagopa-pn-cx-groups                                                                         # NO EXTERNAL
          in: query                                                                               # NO EXTERNAL
          schema:                                                                                 # NO EXTERNAL
            type: array
            items:
              type: string
      responses:                                                                                  # NO EXTERNAL
        '200':                                                                                    # NO EXTERNAL
          description: OK                                                                         # NO EXTERNAL
          content:                                                                                # NO EXTERNAL
            application/json:                                                                     # NO EXTERNAL
              schema:                                                                             # NO EXTERNAL
                type: array                                                                       # NO EXTERNAL
                items:                                                                            # NO EXTERNAL
                  $ref: '#/components/schemas/InternalMandateDto'                                 # NO EXTERNAL
        '400':                                                                                    # NO EXTERNAL
          description: Bad Request                                                                # NO EXTERNAL
          content:                                                                                # NO EXTERNAL
            application/json:                                                                     # NO EXTERNAL
              schema:                                                                             # NO EXTERNAL
                $ref: 'remote-refs.yaml#/components/schemas/Problem'                              # NO EXTERNAL
  /mandate-private/api/v1/mandates-by-internaldelegator/{internaluserId}:                         # NO EXTERNAL
    get:                                                                                          # NO EXTERNAL
      tags:                                                                                       # NO EXTERNAL
        - mandate-private-service                                                                 # NO EXTERNAL
      summary: Return only active mandates for delegator, optionally filtered by single mandateId # NO EXTERNAL
      operationId: listMandatesByDelegator                                                        # NO EXTERNAL
      parameters:
        - name: x-pagopa-pn-cx-type
          in: query
          description: cx-type (PF, PG, PA)
          required: true
          schema:
            $ref: '#/components/schemas/cxTypeAuthFleet'                                        # NO EXTERNAL
        - name: internaluserId                                                                    # NO EXTERNAL
          in: path                                                                                # NO EXTERNAL
          required: true                                                                          # NO EXTERNAL
          schema:                                                                                 # NO EXTERNAL
            type: string                                                                          # NO EXTERNAL
        - name: mandateId                                                                         # NO EXTERNAL
          in: query                                                                               # NO EXTERNAL
          schema:                                                                                 # NO EXTERNAL
            type: string                                                                          # NO EXTERNAL
        - name: x-pagopa-pn-cx-groups                                                                         # NO EXTERNAL
          in: query                                                                               # NO EXTERNAL
          schema:
            type: array
            items:
              type: string
        - name: cx-role                                                                           # NO EXTERNAL
          in: query                                                                               # NO EXTERNAL
          schema:
            type: string
      responses:                                                                                  # NO EXTERNAL
        '200':                                                                                    # NO EXTERNAL
          description: OK                                                                         # NO EXTERNAL
          content:                                                                                # NO EXTERNAL
            application/json:                                                                     # NO EXTERNAL
              schema:                                                                             # NO EXTERNAL
                type: array                                                                       # NO EXTERNAL
                items:                                                                            # NO EXTERNAL
                  $ref: '#/components/schemas/InternalMandateDto'                                 # NO EXTERNAL
        '400':                                                                                    # NO EXTERNAL
          description: Bad Request                                                                # NO EXTERNAL
          content:                                                                                # NO EXTERNAL
            application/json:                                                                     # NO EXTERNAL
              schema:                                                                             # NO EXTERNAL
                $ref: 'remote-refs.yaml#/components/schemas/Problem'                              # NO EXTERNAL
  /mandate/api/v1/count-by-delegate:
    get:
      tags:
        - mandate-service
#      security:                                                                # ONLY EXTERNAL
#        - bearerAuth: [ ]                                                      # ONLY EXTERNAL
      summary: >-
        Return total mandates based on required status if filter's specified. 
        If no filter is present, returns total of all pending and active mandates
      operationId: countMandatesByDelegate
      parameters:
        - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxGroupsAuthFleet'     # NO EXTERNAL
        - $ref: '#/components/parameters/cxRoleAuthFleet'                       # NO EXTERNAL
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MandateCountsDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: 'remote-refs.yaml#/components/schemas/Problem'
  /mandate/api/v1/mandate:
    post:
      tags:
        - mandate-service
      summary: Create a new mandate, returns the just created mandate if successful
      operationId: createMandate
#      security:                                                                # ONLY EXTERNAL
#        - bearerAuth: [ ]                                                      # ONLY EXTERNAL
      parameters:                                                               # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'
        - $ref: 'remote-refs.yaml#/components/parameters/cxGroupsAuthFleet'     # NO EXTERNAL
        - $ref: '#/components/parameters/cxRoleAuthFleet'                       # NO EXTERNAL
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MandateDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MandateDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: 'remote-refs.yaml#/components/schemas/Problem'
  /mandate/api/v1/mandate/{mandateId}/accept:
    patch:
      tags:
        - mandate-service
      summary: Accept a mandate
      operationId: acceptMandate
#      security:                                                                # ONLY EXTERNAL
#        - bearerAuth: [ ]                                                      # ONLY EXTERNAL
      parameters:
        - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxGroupsAuthFleet'     # NO EXTERNAL
        - $ref: '#/components/parameters/cxRoleAuthFleet'                       # NO EXTERNAL
        - name: mandateId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptRequestDto'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: 'remote-refs.yaml#/components/schemas/Problem'
  /mandate/api/v1/mandate/{mandateId}/reject:
    patch:
      tags:
        - mandate-service
      summary: Reject a mandate
      operationId: rejectMandate
#      security:                                                                # ONLY EXTERNAL
#        - bearerAuth: [ ]                                                      # ONLY EXTERNAL
      parameters:
        - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxGroupsAuthFleet'     # NO EXTERNAL
        - $ref: '#/components/parameters/cxRoleAuthFleet'                       # NO EXTERNAL
        - name: mandateId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: 'remote-refs.yaml#/components/schemas/Problem'
  /mandate/api/v1/mandate/{mandateId}/revoke:
    patch:
      tags:
        - mandate-service
      summary: Revoke a mandate
      operationId: revokeMandate
#      security:                                                                # ONLY EXTERNAL
#        - bearerAuth: [ ]                                                      # ONLY EXTERNAL
      parameters:
        - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxGroupsAuthFleet'     # NO EXTERNAL
        - $ref: '#/components/parameters/cxRoleAuthFleet'                       # NO EXTERNAL
        - name: mandateId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: 'remote-refs.yaml#/components/schemas/Problem'
  /mandate/api/v1/mandates-by-delegate:
    get:
      tags:
        - mandate-service
      summary: Return filtered mandates based on required status if filter's specified. If no filter is present, returns all pending and active mandates
      operationId: listMandatesByDelegate_1
#      security:                                                                # ONLY EXTERNAL
#        - bearerAuth: [ ]                                                      # ONLY EXTERNAL
      parameters:
        - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxGroupsAuthFleet'     # NO EXTERNAL
        - $ref: '#/components/parameters/cxRoleAuthFleet'
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MandateDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: 'remote-refs.yaml#/components/schemas/Problem'
  /mandate/api/v1/mandates-by-delegator:
    get:
      tags:
        - mandate-service
      summary: Return all pending and active mandates
      operationId: listMandatesByDelegator_1
#      security:                                                                # ONLY EXTERNAL
#        - bearerAuth: [ ]                                                      # ONLY EXTERNAL
      parameters:                                                               # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxGroupsAuthFleet'     # NO EXTERNAL
        - $ref: '#/components/parameters/cxRoleAuthFleet'                       # NO EXTERNAL
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MandateDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: 'remote-refs.yaml#/components/schemas/Problem'
components:
  parameters:
    cxRoleAuthFleet:
      name: x-pagopa-pn-cx-role
      in: header
      description: Ruolo (estratto da token di Self Care)
      required: false
      schema:
        type: string
  schemas:
    cxTypeAuthFleet:
      type: string
      enum:
        - PG
        - PF
        - PA
    InternalMandateDto:                                                               # NO EXTERNAL
      type: object                                                                    # NO EXTERNAL
      properties:                                                                     # NO EXTERNAL
        mandateId:                                                                    # NO EXTERNAL
          type: string                                                                # NO EXTERNAL
        delegator:                                                                    # NO EXTERNAL
          type: string                                                                # NO EXTERNAL
        delegate:                                                                     # NO EXTERNAL
          type: string                                                                # NO EXTERNAL
        visibilityIds:                                                                # NO EXTERNAL
          type: array                                                                 # NO EXTERNAL
          items:                                                                      # NO EXTERNAL
            type: string                                                              # NO EXTERNAL
        datefrom:                                                                     # NO EXTERNAL
          type: string                                                                # NO EXTERNAL
          description: ISO 8601 format                                                # NO EXTERNAL
        dateto:                                                                       # NO EXTERNAL
          type: string                                                                # NO EXTERNAL
          description: ISO 8601 format                                                # NO EXTERNAL
    MandateCountsDto:
      type: object
      properties:
        value:
          type: integer
          format: int32
    MandateDto:
      type: object
      properties:
        mandateId:
          type: string
          description: Available always, not required on mandate creation.
          nullable: true
        delegator:
          $ref: '#/components/schemas/UserDto'
        delegate:
          $ref: '#/components/schemas/UserDto'
        status:
          type: string
          description: Available always, not required on mandate creation.
          nullable: true
          enum:
            - pending
            - active
            - revoked
            - rejected
            - expired
        visibilityIds:
          type: array
          description: Available always, not required on mandate creation.
          nullable: true
          items:
            $ref: '#/components/schemas/OrganizationIdDto'
        verificationCode:
          type: string
          description: Available only for request where requesting user is the delegator
          nullable: true
        datefrom:
          type: string
          description: ISO 8601 format
        dateto:
          type: string
          description: ISO 8601 format
    OrganizationIdDto:
      type: object
      properties:
        name:
          type: string
        uniqueIdentifier:
          type: string
      description: Available always, not required on mandate creation.
      nullable: true
    UserDto:
      type: object
      properties:
        displayName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        companyName:
          type: string
        fiscalCode:
          type: string
        person:
          type: boolean
      description: Available only for request where requesting user is the delegator
      nullable: true
    AcceptRequestDto:
      type: object
      properties:
        verificationCode:
          type: string
        groups:
          type: array
          items:
            type: string
#  securitySchemes:                                      # ONLY EXTERNAL
#    bearerAuth:                                         # ONLY EXTERNAL
#      type: http                                        # ONLY EXTERNAL
#      scheme: bearer                                    # ONLY EXTERNAL
#      bearerFormat: JWT                                 # ONLY EXTERNAL
#security:                                               # ONLY EXTERNAL
# - bearerAuth: []                                       # ONLY EXTERNAL
              